feat: consolidate line item tests
